feat: add and mkdir functions
